+++ lib/typewritertext.dart
@@ -4,7 +4,14 @@ library typewritertext;
 import 'dart:async';
 import 'package:flutter/material.dart';
 
-enum _TypeWriterTextType { text, builder }
+/// Enum to represent the type of [TypeWriterText].
+enum _TypeWriterTextType {
+  /// Represents a [TypeWriterText] with [Text] content.
+  _text,
+
+  /// Represents a [TypeWriterText] with a custom builder function.
+  _builder,
+}
 
 /// A simple typewriter text animation wrapper for flutter.
 class TypeWriterText extends StatefulWidget {
@@ -18,30 +25,60 @@ class TypeWriterText extends StatefulWidget {
   ///```
   const TypeWriterText({
     Key? key,
-    required this.text,
-    required this.duration,
     this.alignment,
     this.maintainSize = true,
     this.repeat = false,
     this.play = true,
-  })  : _type = _TypeWriterTextType.text,
+    required this.text,
+    required this.duration,
+  })  : _type = _TypeWriterTextType._text,
         builder = null,
+        data = '',
         super(key: key);
 
-  const TypeWriterText.builder({
+  /// Constructor for creating a [TypeWriterText] with a custom builder function.
+  ///
+  /// The typewriter animation will be based on the provided [builder] function,
+  /// which is called on each tick to generate the text to display.
+  ///
+  /// [data] is the text that will be displayed during the typewriter animation.
+  /// [repeat] specifies whether the animation should repeat once completed (default is `false`).
+  /// [play] determines whether the animation should start immediately (default is `true`).
+  /// [duration] defines the time interval between each tick.
+  /// [builder] is a function that generates the text to display based on the current tick.
+  ///
+  /// Example usage:
+  /// ```dart
+  /// const TypeWriterText.builder(
+  ///   "Your text",
+  ///   duration: Duration(milliseconds: 50),
+  ///   builder: (context, value) {
+  ///     return AutoSizeText(
+  ///       value,
+  ///       textAlign: TextAlign.left,
+  ///       style: Theme.of(context).textTheme.titleMedium,
+  ///     );
+  ///   },
+  /// );
+  /// ```
+  const TypeWriterText.builder(
+    this.data, {
     Key? key,
-    this.alignment,
-    this.maintainSize = true,
     this.repeat = false,
     this.play = true,
     required this.duration,
     required this.builder,
-  })  : _type = _TypeWriterTextType.builder,
+  })  : _type = _TypeWriterTextType._builder,
         text = null,
+        alignment = null,
+        maintainSize = false,
         super(key: key);
 
+  /// The text data to be displayed during the typewriter animation.
+  final String data;
+
   /// Widget builder for advance customization.
-  final Widget Function(BuildContext context)? builder;
+  final Widget Function(BuildContext context, String value)? builder;
 
   /// Requires [Text] widget as it's value.
   final Text? text;
@@ -69,15 +106,16 @@ class TypeWriterText extends StatefulWidget {
   ///To set whether animation should be played or not.
   ///
   ///Default value is `true`.
-  final bool? play;
+  final bool play;
 
   @override
   State<TypeWriterText> createState() => _TypeWriterTextState();
 }
 
-class _TypeWriterTextState extends State<TypeWriterText> with SingleTickerProviderStateMixin {
+class _TypeWriterTextState extends State<TypeWriterText>
+    with SingleTickerProviderStateMixin {
   late AnimationController _controller;
-  String _textContent = "";
+  int _tick = 0;
 
   @override
   void initState() {
@@ -87,29 +125,30 @@ class _TypeWriterTextState extends State<TypeWriterText> with SingleTickerProvid
       duration: const Duration(milliseconds: 1),
     );
 
-    if (widget.play == true) {
-      _startTyping(0);
+    if (widget.play) {
+      WidgetsBinding.instance.addPostFrameCallback((timeStamp) {
+        _run();
+      });
     }
   }
 
-  void _startTyping(int index) {
-    if (index < widget.text.data!.length) {
+  void _run() {
+    if (_tick <
+        (widget._type == _TypeWriterTextType._text
+            ? widget.text!.data!.length
+            : widget.data.length)) {
       Future.delayed(widget.duration, () {
         if (mounted) {
-          setState(() {
-            _textContent += widget.text.data![index];
-          });
-          _startTyping(index + 1);
+          setState(() => _tick++);
+          _run();
         }
       });
     } else {
       if (widget.repeat) {
         Future.delayed(widget.duration, () {
           if (mounted) {
-            setState(() {
-              _textContent = "";
-            });
-            _startTyping(0);
+            setState(() => _tick = 0);
+            _run();
           }
         });
       }
@@ -124,50 +163,60 @@ class _TypeWriterTextState extends State<TypeWriterText> with SingleTickerProvid
 
   @override
   Widget build(BuildContext context) {
-    if (widget.play == false) {
-      ///If play is `false`, return original [Text].
-      return widget.text;
-    } else {
-      ///If play is `true`, return animated text.
-      return LayoutBuilder(builder: (_, constraints) {
-        ///Used as a dummy final text so we can get the final width and height.
-        TextPainter textPainter = TextPainter(
-            locale: widget.text.locale,
-            maxLines: widget.text.maxLines,
-            strutStyle: widget.text.strutStyle,
-            text: TextSpan(
-                text: widget.text.data!,
-                style: widget.text.style,
-                locale: widget.text.locale,
-                semanticsLabel: widget.text.semanticsLabel),
-            textAlign: widget.text.textAlign ?? TextAlign.start,
-            textDirection: widget.text.textDirection ?? TextDirection.ltr,
-            textHeightBehavior: widget.text.textHeightBehavior,
-            textScaleFactor: widget.text.textScaleFactor ?? 1.0,
-            textWidthBasis: widget.text.textWidthBasis ?? TextWidthBasis.parent)
-          ..layout(maxWidth: constraints.maxWidth, minWidth: constraints.minWidth);
+    switch (widget._type) {
+      case _TypeWriterTextType._builder:
+        return widget.builder!(context,
+            widget.play ? widget.data.substring(0, _tick) : widget.data);
+      case _TypeWriterTextType._text:
+        if (!widget.play) {
+          // If play is `false`, return original [Text] or [Builder].
+          return widget.text!;
+        } else {
+          ///If play is `true`, return animated text.
+          return LayoutBuilder(builder: (_, constraints) {
+            final text = widget.text!;
 
-        return Container(
-            alignment: widget.alignment,
-            width: widget.maintainSize == true ? textPainter.width : null,
-            height: widget.maintainSize == true ? textPainter.height : null,
-            child: Text(
-              _textContent,
-              key: widget.text.key,
-              locale: widget.text.locale,
-              maxLines: widget.text.maxLines,
-              overflow: widget.text.overflow,
-              semanticsLabel: widget.text.semanticsLabel,
-              softWrap: widget.text.softWrap,
-              strutStyle: widget.text.strutStyle,
-              style: widget.text.style,
-              textAlign: widget.text.textAlign,
-              textDirection: widget.text.textDirection,
-              textHeightBehavior: widget.text.textHeightBehavior,
-              textScaleFactor: widget.text.textScaleFactor,
-              textWidthBasis: widget.text.textWidthBasis,
-            ));
-      });
+            ///Used as a dummy final text so we can get the final width and height.
+            TextPainter textPainter = TextPainter(
+                locale: text.locale,
+                maxLines: text.maxLines,
+                strutStyle: text.strutStyle,
+                text: TextSpan(
+                    text: text.data!,
+                    style: text.style,
+                    locale: text.locale,
+                    semanticsLabel: text.semanticsLabel),
+                textAlign: text.textAlign ?? TextAlign.start,
+                textDirection: text.textDirection ?? TextDirection.ltr,
+                textHeightBehavior: text.textHeightBehavior,
+                textScaleFactor: text.textScaleFactor ?? 1.0,
+                textWidthBasis: text.textWidthBasis ?? TextWidthBasis.parent)
+              ..layout(
+                  maxWidth: constraints.maxWidth,
+                  minWidth: constraints.minWidth);
+
+            return Container(
+                alignment: widget.alignment,
+                width: widget.maintainSize ? textPainter.width : null,
+                height: widget.maintainSize ? textPainter.height : null,
+                child: Text(
+                  text.data!.substring(0, _tick),
+                  key: text.key,
+                  locale: text.locale,
+                  maxLines: text.maxLines,
+                  overflow: text.overflow,
+                  semanticsLabel: text.semanticsLabel,
+                  softWrap: text.softWrap,
+                  strutStyle: text.strutStyle,
+                  style: text.style,
+                  textAlign: text.textAlign,
+                  textDirection: text.textDirection,
+                  textHeightBehavior: text.textHeightBehavior,
+                  textScaleFactor: text.textScaleFactor,
+                  textWidthBasis: text.textWidthBasis,
+                ));
+          });
+        }
     }
   }
 }